10-11-25/speech_synthesizer/requirements.txt
gtts
playsound==1.2.2
